// app/front-end/src/interfaces/IUserAPI.tsx
export default interface IUserAPI {
  name: {
    title: string;
    first: string;
    last: string;
  };
  email: string;
  login: {
    uuid: string;
    username: string;
    password: string;
    salt: string;
    md5: string;
    sha1: string;
    sha256: string;
  };
  dob: {
    date:  string;
    age: number;
  };
  picture: {
    large:  string;
    medium:  string;
    thumbnail:  string;
  };
}